+++ Cryto/src/Maths/Gcd.java
@@ -10,13 +10,13 @@ import java.util.*;
  */
 public class Gcd 
 {
-    int gcd;
+    long gcd;
 	public Gcd()   //gcd of (a,b)=gcd(b,a)
 	{
      
 	}
-   public int calGcd(int n1,int n2)
-   {  int r1=n1,r2=n2,r=0,q=0;
+   public long calGcd(long n1,long n2)
+   {  long r1=n1,r2=n2,r=0,q=0;
 	   while(r2>0)
 	   {
 		   q=r1/r2;
